requirements.txt
tensorflow
boto3
pandas
numpy
matplotlib
notebook
dvc
python-box == 6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scikit-learn
scipy
Flask
Flask-Cors
-e .